requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Array & math
numpy>=1.24.0

# Image I/O & processing
opencv-python>=4.7.0
Pillow>=9.4.0
scikit-image>=0.20.0

# Data handling & utilities
tqdm>=4.65.0

# Plotting
matplotlib>=3.7.0

# Machine learning utilities
scikit-learn>=1.2.0
